Create Vector.h
Templates in C++ are essentially blueprints that the compiler uses to generate code when a specific type is instantiated. The key thing to understand is that template code is not instantiated until the compiler sees a use of the template with a specific type.

When you declare a template class or function in a header file and define it in a source file, the compiler does not generate any code for the template in the source file until it knows what specific types the template will be used with. However, when another source file includes the header and tries to use the template with a specific type, the compiler might not have the necessary information available to generate the required code, leading to a linker error